requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
joblib>=1.3
pyyaml>=6.0
pytest>=7.4
